+++ src/CHANGELOG.md
@@ -15,3 +15,21 @@ Added license
 Improved load performance and prevented unstyled content by waiting for styles before rendering.
 
 ---
+
+## v0.1.3
+
+Improved console outputs. Supported Material Symbols with Multiple Plugins.
+
+---
+
+## v0.1.4
+
+Supports multi-plugins
+
+---
+
+## v0.1.4
+
+Logs improvements
+
+---